refactor patches related to transformers
